+++ ui/admin/app/controllers/scopes/scope/aliases/index.js
@@ -8,16 +8,13 @@ import { inject as service } from '@ember/service';
 import { action } from '@ember/object';
 import { loading } from 'ember-loading';
 import { notifySuccess, notifyError } from 'core/decorators/notify';
-import { tracked } from '@glimmer/tracking';
+import { confirm } from 'core/decorators/confirm';
 
 export default class ScopesScopeAliasesIndexController extends Controller {
   // =services
   @service can;
   @service router;
 
-  // =attributes
-  @tracked showModal = false;
-  @tracked selectedAlias;
   // =actions
 
   /**
@@ -52,24 +49,15 @@ export default class ScopesScopeAliasesIndexController extends Controller {
   }
 
   /**
-   * Handles modal visibility and sets the destination_id to a tracked variable so we can handle the clear button visibility on the modal
-   */
-  @action
-  toggleModal(alias) {
-    this.showModal = !this.showModal;
-    this.selectedAlias = alias;
-  }
-
-  /**
    * Remove destaination_id from alias
    * @param {AliasModel} alias
    */
   @action
   @loading
+  @confirm('questions.clear-confirm')
   @notifyError(({ message }) => message, { catch: true })
   @notifySuccess('notifications.clear-success')
   async clearAlias(alias) {
-    this.showModal = false;
     alias.destination_id = '';
     await alias.save();
     await this.router.refresh();
@@ -81,11 +69,11 @@ export default class ScopesScopeAliasesIndexController extends Controller {
    */
   @action
   @loading
+  @confirm('resources.alias.messages.delete')
   @notifyError(({ message }) => message, { catch: true })
   @notifySuccess('notifications.delete-success')
   async deleteAlias(alias) {
     await alias.destroyRecord();
-    this.showModal = false;
     await this.router.replaceWith('scopes.scope.aliases');
     await this.router.refresh();
   }
