+++ ui/admin/tests/acceptance/aliases/update-test.js
@@ -7,12 +7,16 @@ import { module, test } from 'qunit';
 import { visit, click, fillIn } from '@ember/test-helpers';
 import { setupApplicationTest } from 'admin/tests/helpers';
 import setupMirage from 'ember-cli-mirage/test-support/setup-mirage';
+import { setupIndexedDb } from 'api/test-support/helpers/indexed-db';
 
 import { authenticateSession } from 'ember-simple-auth/test-support';
 
 module('Acceptance | aliases | update', function (hooks) {
   setupApplicationTest(hooks);
   setupMirage(hooks);
+  setupIndexedDb(hooks);
+
+  let aliasCount;
 
   const SAVE_BTN_SELECTOR = '.rose-form-actions [type="submit"]';
 
@@ -21,6 +25,11 @@ module('Acceptance | aliases | update', function (hooks) {
   const NAME_FIELD_TEXT = 'random string';
 
   const BUTTON_SELECTOR = '.rose-form-actions [type="button"]';
+  const CLEAR_DROPDOWN_SELECTOR =
+    '.hds-dropdown-list-item--color-action:nth-child(2) button';
+
+  const DROPDOWN_BUTTON_SELECTOR = '.hds-dropdown-toggle-icon';
+  const DROPDOWN_ITEM_SELECTOR = '.hds-dropdown-list-item';
 
   const instances = {
     scopes: {
@@ -50,6 +59,7 @@ module('Acceptance | aliases | update', function (hooks) {
     urls.globalScope = `/scopes/global`;
     urls.aliases = `${urls.globalScope}/aliases`;
     urls.alias = `${urls.aliases}/${instances.alias.id}`;
+    aliasCount = () => this.server.schema.aliases.all().models.length;
 
     authenticateSession({});
   });
@@ -77,5 +87,22 @@ module('Acceptance | aliases | update', function (hooks) {
 
     assert.dom(NAME_FIELD_SELECTOR).hasValue(`${name}`);
     assert.strictEqual(instances.alias.name, name);
+  });
+
+  test('users have the option to clear an alias', async function (assert) {
+    const count = aliasCount();
+    assert.true(instances.alias.authorized_actions.includes('read'));
+
+    await visit(urls.globalScope);
+
+    urls.alias = `${urls.aliases}/${instances.alias.id}`;
+
+    await click(`[href="${urls.aliases}"]`);
+    await click(DROPDOWN_BUTTON_SELECTOR);
+
+    assert.dom(DROPDOWN_ITEM_SELECTOR).exists();
+    assert.dom(CLEAR_DROPDOWN_SELECTOR).exists();
+    await click(CLEAR_DROPDOWN_SELECTOR);
+    assert.strictEqual(aliasCount(), count);
   });
 });
